refactor: Consolidate command files
This commit refactors the command file structure by merging related commands into single files, as requested. This improves organization and reduces the number of files in the project.

The following refactors were performed:
- `tpastatus.js` was merged into `tpa.js`.
- The consolidated bounty file was renamed from `rbounty.js` to `bounty.js`.
- `baltop.js` was merged into `balance.js`.

The command loader (`index.js`) has been updated to reflect these changes, and all redundant files have been deleted.
