+++ AddonExeBP/scripts/modules/commands/tpa.js
@@ -1,7 +1,8 @@
 import { commandManager } from './commandManager.js';
-import * as tpaManager from '../../core/tpaManager.js';
+import { createRequest, acceptRequest, denyRequest, cancelRequest, getOutgoingRequest, getIncomingRequest } from '../../core/tpaManager.js';
 import { getConfig } from '../../core/configManager.js';
 import { getCooldown, setCooldown } from '../../core/cooldownManager.js';
+import { playSound } from '../../core/utils.js';
 
 commandManager.register({
     name: 'tpa',
@@ -38,7 +39,7 @@ commandManager.register({
             return;
         }
 
-        const result = tpaManager.createRequest(player, targetPlayer, 'tpa');
+        const result = createRequest(player, targetPlayer, 'tpa');
 
         if (result.success) {
             setCooldown(player, 'tpa');
@@ -79,7 +80,7 @@ commandManager.register({
             return;
         }
 
-        const result = tpaManager.createRequest(player, targetPlayer, 'tpahere');
+        const result = createRequest(player, targetPlayer, 'tpahere');
 
         if (result.success) {
             player.sendMessage(`§aTPA Here request sent to ${targetPlayer.name}. They have ${config.tpa.requestTimeoutSeconds} seconds to accept.`);
@@ -104,7 +105,7 @@ commandManager.register({
             return;
         }
 
-        tpaManager.acceptRequest(player);
+        acceptRequest(player);
     }
 });
 
@@ -122,7 +123,7 @@ commandManager.register({
             return;
         }
 
-        tpaManager.denyRequest(player);
+        denyRequest(player);
     }
 });
 
@@ -139,6 +140,42 @@ commandManager.register({
             return;
         }
 
-        tpaManager.cancelRequest(player);
+        cancelRequest(player);
+    }
+});
+
+commandManager.register({
+    name: 'tpastatus',
+    description: 'Checks the status of your outgoing and incoming TPA requests.',
+    category: 'TPA System',
+    permissionLevel: 1024, // Everyone
+    parameters: [],
+    execute: (player, args) => {
+        const outgoing = getOutgoingRequest(player);
+        const incoming = getIncomingRequest(player);
+
+        let statusMessage = '§a--- TPA Status ---\n';
+        let foundRequest = false;
+
+        if (outgoing) {
+            foundRequest = true;
+            const typeText = outgoing.type === 'tpa' ? 'teleport to them' : 'teleport them to you';
+            statusMessage += `§eOutgoing Request:§r You have sent a request to §b${outgoing.targetPlayerName}§r to ${typeText}.\n`;
+            statusMessage += '§7(Use /tpacancel to cancel this request)\n';
+        }
+
+        if (incoming) {
+            foundRequest = true;
+            const typeText = incoming.type === 'tpa' ? 'teleport to you' : 'teleport you to them';
+            statusMessage += `§eIncoming Request:§r You have a request from §b${incoming.sourcePlayerName}§r to ${typeText}.\n`;
+            statusMessage += '§7(Use /tpaccept or /tpadeny to respond)\n';
+        }
+
+        if (!foundRequest) {
+            statusMessage += '§fYou have no pending TPA requests.';
+        }
+
+        player.sendMessage(statusMessage.trim());
+        playSound(player, 'random.orb');
     }
 });
